pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "culpa"
version = "0.1.0"
description = "Deterministic replay & counterfactual debugging for AI coding agents"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
keywords = ["ai", "agents", "debugging", "replay", "llm", "observability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "httpx>=0.27",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "python-multipart>=0.0.9",
    "stripe>=8.0",
    "resend>=2.0",
]
anthropic = [
    "anthropic>=0.30",
]
openai = [
    "openai>=1.30",
]
litellm = [
    "litellm>=1.40",
]
watch = [
    "watchdog>=4.0",
]
all = [
    "culpa[server,anthropic,openai,litellm,watch]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-aiohttp>=1.0",
    "httpx>=0.27",
    "mypy>=1.10",
    "ruff>=0.4",
]

[project.scripts]
culpa = "culpa.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["sdk/culpa"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
